Update Azure VM manager and add new packages
Updated the Azure Virtual Machine Manager to use the Fluent library for Azure management. This simplifies the setup and configuration process for Azure resources. Also added two new package references for Compute and Management Fluent in the AzureProject.csproj file which will be used to manage Azure resources in a more streamlined manner.